requirements.txt
flask
mysql-connector-python
werkzeug
qrcode[pil]
python-dotenv
gunicorn